ml_project/requirements.txt
py==1.10.0
Faker==8.0.0
marshmallow_dataclass==8.4.1
pandas==1.2.4
dataclasses==0.6
pytest==6.2.3
click==7.1.2
numpy==1.20.1
PyYAML==5.4.1
scikit_learn==0.24.2
